add opt request

+++ src/main/java/com/seckillmall/service/model/UserModel.java
@@ -5,7 +5,7 @@ public class UserModel {
 
     private String name;
 
-    private Integer gender;
+    private Byte gender;
 
     private Integer age;
 
@@ -41,11 +41,11 @@ public class UserModel {
         this.name = name;
     }
 
-    public Integer getGender() {
+    public Byte getGender() {
         return gender;
     }
 
-    public void setGender(Integer gender) {
+    public void setGender(Byte gender) {
         this.gender = gender;
     }
